+++ src/Symfony/Component/Messenger/Tests/Middleware/SendMessageMiddlewareTest.php
@@ -31,7 +31,7 @@ class SendMessageMiddlewareTest extends MiddlewareTestCase
 
         $middleware = new SendMessageMiddleware(new InMemorySenderLocator($sender));
 
-        $sender->expects($this->once())->method('send')->with($envelope);
+        $sender->expects($this->once())->method('send')->with($envelope)->willReturn($envelope);
 
         $middleware->handle($envelope, $this->getStackMock(false));
     }
@@ -43,7 +43,7 @@ class SendMessageMiddlewareTest extends MiddlewareTestCase
 
         $middleware = new SendMessageMiddleware(new InMemorySenderLocator($sender));
 
-        $sender->expects($this->once())->method('send')->with($envelope);
+        $sender->expects($this->once())->method('send')->with($envelope)->willReturn($envelope);
 
         $middleware->handle($envelope, $this->getStackMock(false));
     }
@@ -58,7 +58,7 @@ class SendMessageMiddlewareTest extends MiddlewareTestCase
             DummyMessage::class => true,
         ));
 
-        $sender->expects($this->once())->method('send')->with($envelope);
+        $sender->expects($this->once())->method('send')->with($envelope)->willReturn($envelope);
 
         $middleware->handle($envelope, $this->getStackMock());
     }
@@ -73,7 +73,7 @@ class SendMessageMiddlewareTest extends MiddlewareTestCase
             DummyMessage::class => true,
         ));
 
-        $sender->expects($this->once())->method('send')->with($envelope);
+        $sender->expects($this->once())->method('send')->with($envelope)->willReturn($envelope);
 
         $middleware->handle($envelope, $this->getStackMock());
     }
@@ -88,7 +88,7 @@ class SendMessageMiddlewareTest extends MiddlewareTestCase
             DummyMessageInterface::class => true,
         ));
 
-        $sender->expects($this->once())->method('send')->with($envelope);
+        $sender->expects($this->once())->method('send')->with($envelope)->willReturn($envelope);
 
         $middleware->handle($envelope, $this->getStackMock());
     }
@@ -103,7 +103,7 @@ class SendMessageMiddlewareTest extends MiddlewareTestCase
             '*' => true,
         ));
 
-        $sender->expects($this->once())->method('send')->with($envelope);
+        $sender->expects($this->once())->method('send')->with($envelope)->willReturn($envelope);
 
         $middleware->handle($envelope, $this->getStackMock());
     }